+++ forkbuntu/services/__init__.py
@@ -2,9 +2,9 @@ from .cache import Cache
 from .clean import Clean
 from .configure import Configure
 from .extras import Extras
+from .filesystem import Filesystem
 from .gpg import GPG
 from .initrd import Initrd
-from .pack import Pack
+from .iso import Iso
 from .setup import Setup
-from .unpack import Unpack
 from .util import Util