+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.2.3] - 2018-06-01
 ### Added
 * modify initrd filesystem
+### Changed
+* service organization
 ### Fixed
 * empty checksum bug